requirements.txt
opencv-python>=4.9.0.80
scenedetect[opencv]>=0.6.4
numpy>=1.26.4
Pillow>=10.3.0

# 安装提示：使用以下命令安装依赖
# pip install -r requirements.txt --only-binary :all:
